Split `assert_integer_form` utility according to its two distinct purposes.
One purpose, more common in the code base, is to assert that incoming
samples of integer-support distributions are, in fact, integers.  It
is in general not necessary to ensure that such samples are
representable as (32-bit) integers if the downstream code will not try
to cast them.

This purpose is now served by a changed `assert_integer_form` that
tests integerness by rounding rather than casting, thus admitting
floats greater than 2**31 as integers.  It also permits a tolerance
parameter, so that, for example, tf.math.log(tf.math.exp(1.))
registers as an integer despite evaluating to 0.99999994 in 32-bit
precision.

The other, slightly less common purpose is to guard distribution
parameters that the implementation _was_ going to cast to an integer
type for one or another reason.  These cases are dominated by `count`
parameters of binomial-like distributions.  This purpose is now served
by a new `assert_casting_closed` helper, which does what the old
`assert_integer_form` helper used to do.

This change also surveys uses of `assert_integer_form` in the code
base and adjusts those that actually require the casting closure check
to request it.

PiperOrigin-RevId: 335686716