Delete the `use_static_graph=False` sampler from Mixture.
Leave the constructor flag, but mark it deprecated and document that
it does nothing.

Why?
- It had a bug causing quadratic scaling in the batch size.
- Code complexity.
- Not compatible with XLA and TPUs.

For the record, the advantage of the now-deleted code path, relative
to `use_static_graph=True`, was that it could sometimes do less
sampling of the component distributions.  Specifically, a batch-1
Mixture would only sample from the sub-component that was actually
selected, and not from the others.  More precisely, as written the
code would build TF graphs to sample from each component, but it would
ask all but one of them for zero samples at graph execution time.

This advantage, such as it is, inevitably degrades as the batch size
increases: In a batch of 1000 4-way Mixtures with relatively even
mixing weights, every component would surely have to produce a sample
in one or another batch member, so there is nothing to gain by trying
to avoid those samples.  The bug was that instead of stopping at 1
sample per component, the code path would ask for ~250 each.
PiperOrigin-RevId: 336893201