Pass the momentum_distribution from preconditioned HMC in kernel results.
This will allow preconditioned HMC to read a momentum distribution from the kernel results, with will provide for a transition kernel that learns and updates the momentum. It relies on the momentum distribution being a composite tensor whenever `store_parameters_in_results` is True (which is implicitly true whenever there is a step size adapting kernel).

PiperOrigin-RevId: 337131982