Normalize kernel results seen by `Reducer`s in `sample_fold`.
Specifically, `Reducer`s passed to `sample_fold` should see the kernel
results of the kernel passed to `sample_fold`, not of the
`SampleDiscardingKernel` that `sample_fold` itself introduces.

To this end, add a kernel results adjustment function to
`WithReductions`, and use it in `sample_fold` to strip the layer of
kernel results wrapping introduced by `SampleDiscardingKernel`.

PiperOrigin-RevId: 337146314