Doc update to Sample class.
One of my misinterpretations of Sample was that it makes an empirical distribution
by (i) sampling from a distribution, (ii) freezing these samples and making a
static empirical distribution. It is not.

Another misinterpretation I had of Sample was that the samples would be placed
in the leftmost dimension, as is done elsewhere in TFP.

Another misinterpretation I had is that this is a "sampling distribution" (https://en.wikipedia.org/wiki/Sampling_distribution).  It is not, although the initial docstring line about "reducing over..." hinted in that direction.

There was also a bug in the docs regarding a reduction axis.

PiperOrigin-RevId: 338448353