Rename `interpolate_nondiscrete` flag to `force_probs_to_zero_outside_support` (with the opposite sense) in Zipf.
The second name leaves room for other distributions to use the same
flag to control tf.where gates on extrapolating the support to, e.g.,
negative arguments.

Not changing the pre-existing tf.where gate in Zipf.cdf to actually be
controlled by said new flag, because that would change the behavior of
Zipf with no flags set.  That can be done another time.

PiperOrigin-RevId: 340878676