Rename `interpolate_nondiscrete` flag to `force_probs_to_zero_outside_support` (with the opposite sense) in Skellam.
The second name leaves room for other distributions to use the same
flag to control tf.where gates on extrapolating the support to, e.g.,
negative arguments.

Not going through the usual deprecation process because Skellam was
only contributed a few days ago, and in particular was not in any
stable release yet.

PiperOrigin-RevId: 341064310