Add Chandrupatla's method for finding roots of scalar functions.
It's not clear that we *really* need this, but I got most of the way through implementing it before I noticed that `secant_root` already existed. :-) So I figured I might as well check it in, and it does has somewhat different properties.

I tried to mostly mimic the interface of `secant_root`. I think the main practical distinction is that Chandrupatla's method takes a pair of bounds, instead of just an initial position, and is guaranteed to converge if those bounds bracket a root.

PiperOrigin-RevId: 341901341