Add transition kernel for online adaptation of a diagonal mass matrix during MCMC.
The `PreconditionedHamiltonianMonteCarlo` distribution accepts a momentum distribution that it uses to sample momentum from. For example, if you suspect your posterior covariance is `S`, then using a kinetic energy function of `K(p) = 0.5 * (p.T S p)` may improve the performance of Hamiltonian Monte Carlo.

This transition kernel is intended to use to estimate a diagonal covariance matrix, by updating the underlying momentum distribution according to an online estimate of the posterior variance. Note that this is *not* a valid MCMC scheme, but the user may use the estimate this transition kernel produces to supply a fixed momentum distribution for sampling.

PiperOrigin-RevId: 343530087