Enable use of `auto_composite_tensor` without altering extra class hierarchy.
The solution is not beautiful. In order to construct a valid CT subclass *class
instance* to pass to the decorator, the subclass needs to have a _type_spec
property (else, ABCMeta complains and class creation fails, before the
decorator can ever run). So we're forced to implant a trivial _type_spec
definition, which is then overwritten by the decorator.

This change also refactors the internal logic of the auto_composite_tensor,
somewhat substantially, to separate a bit more the AutoCompositeTensorTypeSpec
behaviors from the decorator.

This change also uses the new behavior in `tfp.experimental.stats`.

PiperOrigin-RevId: 343575343