Fix ildj gradients in ScalarFunctionWithInferredInverse bijector.
When defining the inverse log-det-jacobian in terms of an automatic derivative, as we do here, its gradient requires *second* derivatives of the underlying inverse function. We therefore need to provide annotations for the second derivatives, using the inverse function theorem.

It's a bit of a mess to do this in a way that works for both TF and JAX, but it seems to have worked. I'd welcome suggestions for making this neater.

PiperOrigin-RevId: 343928308