Allow using arbitrary structures for make_sharded_log_prob_parts.
The real issue is illustrated by the test added to joint_distribution_test, where
prior to this fix you couldn't actually compute gradients with respect to the
samples of joint distributions due to some structure mismaches. The fix is
likely extremely overkill, but to the extent that the documentation for
`make_sharded_log_prob_parts` doesn't single out lists/tuples as the only
supported structure, this seems an acceptable thing to do.

PiperOrigin-RevId: 345327288