Fix error in preconditioned HMC.
In situations where the momentum distribution was stored in the kernel results, the kinetic energy function passed to the leapfrog integrator would be the default (Gaussian, using an identity covariance matrix), leading to bad results. This, in particular, will fix the behavior of DiagonalMassMatrixAdaptation, which was hitting this code path.

PiperOrigin-RevId: 354133070