Make deprecation warning clearer
The message to 'combine scale_identity_multiplier with scale_diag' seemed to suggest that the arguments `scale_identity_multiplier` and `scale_diag` should be combined together, but what is actually meant is that `scale_diag` should be used in favour of `scale_identity_multiplier`. The comment down below in the body of `__init__` was clearer.