Refactor ASVI to support nested joint- and meta-distributions.
This CL destructures ASVI into three main methods, which can call each other recursively:
 - _asvi_surrogate_for_distribution: handles meta-distributions and dispatches to the other methods.
 - _asvi_surrogate_for_joint_distribution: creates a JDC surrogate that loops over  components of a JD model.
 - _asvi_convex_update_for_base_distribution: creates a convex-update surrogate for a leaf node or 'base' distribution.

This allows the leaf code to be swapped out without disrupting the larger structure.

It also removes separate variable creation (`_make_asvi_trainable_variables`), in favor of a style where variables are created inline as the surrogate is built. The structure of the variables mirrors the structure of the surrogate (e.g., a nested joint distribution will need its own set of nested variables), so building the variables 'locally' avoids the need to keep a variable-building method and a surrogate-building method in sync with duplicate control flow.

All of the private ASVI methods take and return a `variables` arg: if this is `None`, new variables are built and returned; otherwise the surrogate is constructed with the passed-in variables.

PiperOrigin-RevId: 363502109