Deprecate Distribution subclasses inheriting `_parameter_properties`.
Generally, subclasses don't have the same parameters as their parents. A prototypical case is the relationship between TransformedDistribution, which has parameters `distribution` and `bijector`, and the MVN distributions, which have parameters like `loc` and `scale`. The question is: if I define a new subclass

class MyMVN(TransformedDistribution):

  def __init__(self, loc, scale):
    parameters = dict(locals())
    super(MyMVN, self).__init__(
      distribution=..., bijector=..., parameters=parameters)

that doesn't implement _parameter_properties, then what should `MyMVN.parameter_properties()` return?

Default behavior: it would return TransformedDistribution.parameter_properties() (i.e.,  `{'distribution': ..., 'bijector': ...}`) which seems incorrect in that it doesn't correspond to the `self.parameters` of a MyMVN instance.

Behavior implemented by this CL: it will raise NotImplementedError. This makes sense to me since there's no reason in general to assume that MyMVN accepts the same parameters as its parent.

As an exception, it is safe to inherit parameter_properties for lightweight wrapper
classes that pass through directly to the parent __init__. Specifically, inheritance is allowed if:
 (a) The subclass doesn't define __init__, or
 (b) The subclass defines `__init__(self, *args, **kwargs)`.
In the latter case, we assume that parameters are passed through to the parent class. This assumption is not enforceable (one could write a class that parses *args/**kwargs to define its own arguments) and somewhat brittle (taking an explicit 'name' or 'validate_args' prevents inheritance) but seems generally reasonable.

PiperOrigin-RevId: 365142640