Miscellaneous fixes to several distributions.
These are required for a future change adding batch shape inference (cl/350261091), but I've pulled this out as a separate CL because these changes are tangential to the main purpose of that change.

Much of this change is to add ParameterProperties to RelaxedOneHotCategorical and SinhArcsinh, and disable some failing hypothesis tests (that were not previously being run). These explicit properties replace those that these distributions would otherwise (in the future) inherit from TransformedDistribution.

PiperOrigin-RevId: 365693255