Allow custom distribution substitutions in ASVI, and add a few defaults.
I think custom substitutions make the ASVI code more flexible, and in particular, they're helpful in cases where you want to exploit graph structure even though you know that the prior family itself would be a bad surrogate.

For example, the hierarchical representation of a Horseshoe sparse regression model:

tau ~ HalfCauchy(loc=0., scale=1.)
lamb ~ HalfCauchy(loc=0., scale=ones([ndims]))
weights ~ Normal(loc=0., scale=sqrt(tau * lamb))

has latents with HalfCauchy priors, but they parameterize the variance of a Normal, so the posteriors are probably closer to InverseGamma. Even if we didn't believe this, we might want to avoid the Cauchy surrogate in any case because the heavy tails make it difficult to estimate the ELBO. There's still dependence in this model, though, so ASVI should be structurally helpful even if we need to adjust the distribution families.

This change also tries a bit harder to build a surrogate that has the exact same structure as the model (when the model is a JDC), so that we can avoid a Restructure bijector. This makes it easier to test how distributions are being substituted, and just requires being careful to construct each surrogate distribution with the same `name` parameter as the corresponding distribution in the model.

PiperOrigin-RevId: 365709196