Invoke autobatching for JDPinned support bijectors when required.
Previously, the support bijector for a JDPinned distribution was not autobatched, even when the underlying model required autobatching.

It turns out that this is somewhat subtle to fix. It's not sufficient for JDPinned to construct a standard _DefaultJointBijectorAutoBatched instance, because JDPinned exposes pinned values at the lowest level (from `JDPinned._model_coroutine`); so they'd get accessed directly from *inside* the vectorized functions. This breaks the underlying model in general when the pinned values have batch dimensions. Instead, we need to pass the pinned values as *inputs* to vectorized functions, just as we do the unpinned values, so that the model can vectorize over their batch dimensions.

This CL slightly modifies the _DefaultJointBijectorAutoBatched to
1. Use composition (self._joint_bijector) rather than inheriting from _DefaultJointBijector.
2. Factor out the mechanism that vectorizes each member function into a method `self._vectorize_member_fn`, which can be overridden.

It then defines a subclass _DefaultJointBijectorAutoBatchedWithPins to
address the issue above, by defining vectorized bijector member functions that take the pinned values as inputs.

PiperOrigin-RevId: 366144985