Adjust IncrementLogProb and Sharded in order to work together.
Sharded calls the property `experimental_shard_axis_names`, so we add it to IncrementLogProb.  Also, at some point Sharded ends up calling `sample` with a `sample_shape` which is a scalar, ie 0-dimensional tensor, so we reshape that.  Finally, `IncrementLogProb` intentionally does not have a `log_prob` method, so in order for it to work with `JointDensityCoroutine`, we need to expose `unnormalized_log_prob` in Sharded.

PiperOrigin-RevId: 368623920