Handle inputs with sample shape `[1]`, `[1, 1]`, etc. in autobatched JointDistributions.
The existing logic tries to be smart about noticing when the inputs don't actually require autobatching, but it isn't smart enough---it will naively run the model with the given inputs, and this can break things when the rank is higher than the model expects.

The fix is to flatten size-1 sample shapes to `[]` before running the model, and then unflatten them afterwards.

PiperOrigin-RevId: 369993295