Use unconstrained 0 instead of pulling back a sample as a "nice" unconstrained value in Masked.
Why?  Because a pulled-back sample is not reliably "nice".  For
instance, the added test failed before this change, because one of the
samples would pull back to an unconstrained vector with `nan` entries.

However, 0 in unconstrained space should always be "nice", as the idea
is that all of unconstrained R^n is legit.

PiperOrigin-RevId: 371330546