Standardize handling of **kwargs and values in joint distributions.
This is a refactor, not intended to be user-visible except insofar as it results in more consistent error checking. It includes the following changes:

- Moves the `cast_structure` utility from distribution.py to nest_util.py, so we can use it in joint distributions.

- Adds a `_sanitize_value` method to ensure all value elements are Tensors (or None) and that the value has structure matching `self.dtype`. This replaces the `_pad_value_to_full_length` method in jd_vmap_mixin.

- Generalizes the existing method `_resolve_value_from_kwargs` into `_resolve_value`, which handles the case where `value` is itself an arg or kwarg.

- Sets the convention that `_resolve_value` is called by the outermost JD methods: internal methods never see or handle **kwargs.

This likely fixes some theoretical bugs in methods like `_sample_n` that checked `value is None` before the **kwargs had been incorporated. It also likely fixes bugs resulting from the fact that `sample` and `sample_distributions` didn't always do the same input validation.

A consequence of this change is that passing unrecognized arguments to `log_prob` methods now raises a ValueError (consistent with the `sample` methods) rather than a TypeError (as you'd get from calling an ordinary callable with unrecognized kwargs). I don't think this breaks any explicit contract, though it does require some changes to our tests.

PiperOrigin-RevId: 371823551