Robustify preconditioned NUTS tests by weakening the ESS condition.
The current condition `assertAllClose(ess, 100)` requires perfect-or-better mixing across all variables, which is possible (a well-tuned sampler can have negative autocorrelation) but doesn't seem to occur reliably across random seeds. In particular, changes that do nothing but change the effective random seed can break these tests.

This change relaxes the minimum ESS requirement to 50, which maybe we could call 'decent' mixing. I'm not sure if this is the optimal test condition, but the failure that's been blocking cl/372194361 had a minimum ESS of 53, so this is at least sufficient to unblock things there.

PiperOrigin-RevId: 373003922