Generalize JointDistribution `_execute_model` to accept a `sample_and_trace_fn`.
This allows tracing of arbitrary quantities in addition to (or instead of) just distributions and sampled values. For example, we can trace:
 - sample_and_log_prob
 - just samples (to avoid returning dists from inside a vmap)
 - just log_prob, or other map_measure_over_dists quantities --- again allowing us to avoid returning dists from inside a vmap.

(in the near future, we will actually be able to return dists from vmap in most cases, but it still seems nicer to avoid invoking CompositeTensor machinery when it's not necessary).

The motivation for this change is as preparation to support an `experimental_sample_and_log_prob` method, which will be added in a future change.

This also allows for consolidation of redundant code across methods (eg _sample_n and _sample_distributions). I've gone ahead and defined the property `use_vectorized_map` for all JDs (for non-autobatched JDs, it is always False), which makes it possible to strip down `joint_distribution_vmap_mixin` to the bare essentials.

PiperOrigin-RevId: 373054761