Make tfd.Sharded return a tfb.Sharded-wrapped default event space bijector.
This alone enables the simple/common case of
sharded_jd.default_event_space_bijector() to work, where the componentwise
event bijectors do not depend on global parameters. When this assumption is
violated, wrong results will still be returned. Fixing this requires a notion
of a sharded multipart bijector, akin to the sharded JD's.

The broken case can be manually worked-around by a careful use of pbroadcast, see tests.

PiperOrigin-RevId: 373882773