Define experimental batch shape methods for Bijectors.
This opens the door to correctly calculating `log_det_jacobians` via autodiff (the current implementation is incorrect for bijectors with batch shape), automatically broadcasting the batch shapes of TransformedDistributions, batch slicing for TransformedDistributions, and returning Bijectors and TransformedDistributions from vectorized_map calls, among possibly other things.

Q: Why do bijector batch shapes depend on event_ndims?
A: One way to think about this is that batch shape is (up to broadcasting) the shape of the LDJ. Since LDJs are conditional on `event_ndims`, it's not surprising that batch shape would be too.

We *could* try to define a batch shape inherent to each bijector, but JointMap (and other bijectors with `parts_are_independent==True`) makes this hard because the different parts can have different batch shapes. In the absence of a concrete event_ndims to tell us how the parts align, the only batch shape we could define would be structured. It turns out that working with structured batch shapes is a pretty big pain (eg, it puts some heavy demands on TransformedDistribution's parameter_properties.event_ndims annotations). After going down this road for a while, it seemed a lot simpler to live in a world where the batch shape is always just a single shape.

PiperOrigin-RevId: 374960009