Add an `experimental_sample_and_log_prob` method to TFP Distributions.
This is an alternative to bijector caching, intended to be useful (at least) in cases where the cache would otherwise miss. For example, autobatched JDs break caching (unavoidably?), because the values that the bijector sees inside the vmap are from a different graph than those returned to the user. But `sample_and_log_prob` still works in this case for efficient VI.

Specialized implementations for specific distributions will be added in a future change.

PiperOrigin-RevId: 374990765