Minor cleanup of JD*AB to keep `iid_sample` out of stack traces for log_prob.
I found it confusing to see `iid_sample` in the stack trace for a deterministic `log_prob` call (since we now use the same code for everything). Performance-wise, `iid_sample` was already cheap when sample_shape=(), so this won't be a big optimization but also shouldn't make things any worse.

PiperOrigin-RevId: 376042972