Allow partially specifying values for nested joint distributions.
The issue was twofold. First, the various sample_and_value functions were over-eager in bypassing the sample method. This was fixed via a cludge-like change, it kind of feels like the `value` kwarg should be moved up to Distribution so we can just call sample unconditionally.

The second issue was two places where we used map_structure to convert value to a nest of Tensors. This isn't a reasonable thing to do when there are None's present. This was replaced with a recursive call of a utility implemented via _model_flatten/unflatten etc.

PiperOrigin-RevId: 377168479