Make interacting default event space bijectors work for sharded JDs in simple cases.
The cases where it should work where all RVs are tensor-valued. This excludes
nested JDs as well as JDs with non-JD multipart components (which we can't express anyway).

A nontrivial change here is that _sanitize_value was moved into
sample_and_trace functions because using _execute_model inside the bijector
like I did opens up the possibility of elements of the `value` arg to
`_execute_model` to be of a different structure than the actual distribution.
This feels like a good thing anyway, if we interpretat the `_execute_model` as
a primitive effect system.

Ideally I'd make the base JD's use the new _conditioned_bijectors function, but doing so interacted poorly with autobatched JDs for unexplored reasons.

PiperOrigin-RevId: 379403363