Make tf-nightly version-finding script for tests more robust.
Before, we just looked for any releases that weren't in the BAD_DATES list.
This would fail to recognize that some releases don't have a package for the
particular python version being used in the test env.

With this change, we can now be python_version-aware when looing for a not-bad
tf-nightly build. I went ahead and moved a bit more logic into the python
script, to simplify the bash portion. The inline python script-in-a-string is
getting out of hand. I filed a bug to clean this up.

PiperOrigin-RevId: 381392354