Tweak STS default priors to better handle constant series.
This fixes an issue where the heuristic prior for scale parameters, which is typically based on the observed stddev, would put all of its mass on zero when the observed stddev is zero. (we would expect the *posterior* to concentrate near zero, but that's no reason to force the prior to be so informative).

PiperOrigin-RevId: 383501234