Enforce Distribution statistics' batch shape contract by broadcasting.
The contract is that a Distribution object is supposed to look like
all of its parameters' batch shapes are broadcasted together to form
the batch shape of the Distribution.

This mostly happens automatically, but some statistics don't actually
depend on all of the respective Distribution's parameters.  (The most
common case is the variance of a location-scale family doesn't depend
on the location.)  This CL inserts tf.broadcast_to at the outputs of
appropriate statistic methods to enforce the contract, and updates a
Hypothesis test to check for shape-correctness of instantiable
Distributions mechanically.

The Distribution methods affected are
- GeneralizedPareto.entropy
- PERT.mode
- Triangular.entropy
- LambertWNormal.{variance,stddev}
- ExponentiallyModifiedGaussian.{mean,variance,stddev}

PiperOrigin-RevId: 383723788