BREAKING CHANGE: Stop unpacking seeds when splitting in JAX
Before this change `tfp.random.split_seed` would return a list of seeds instead.
of an array of seeds, causing a gather for each seed. With large number of seeds,
this can cause a slowdown in both trace and compile time. This change returns
an array of seeds. If your code relies on using a list of seeds instead of an
array, you can wrap the call to `tfp.random.split_seed` in a `list` or
`jnp.unstack`.

PiperOrigin-RevId: 386558360