Fix an embarassing off-by-one error in STS anomaly detection predictive distributions.
I'd understood `tfp.sts.one_step_predictive` to return a distribution over timesteps `1:T + 1`, so the previous code went to a lot of effort to send it a series padded with an extra initial step in order to get results for steps `0:T` (matching the Gibbs predictive dist). But it turns out that I was wrong, and this padding is not only unnecessary but actually screws things up.

This change also fixes a dtype issue with the Gibbs predictive distribution that came up in testing the above fix.

PiperOrigin-RevId: 388789203