Add custom log-prob ratio to IncrementLogProb.
A typical use case of IncrementLogProb is to implement the likelihood term which commonly has many independent components. Normally, you could use tfd.Independent to take advantage of this to compute log-prob ratios accurately, but that's not always an option for general likelihoods.

Along the way, added support for callable log_prob_increment and refactored
some methods.

PiperOrigin-RevId: 391117861