Define some tangent spaces.
This change sets up the framework for tracking base measures and computing corrections on transformed densities. We introduce a `TangentSpace` base class that defines methods for special case transformations, such as those that are dimension-preserving or coordinate-wise. We also define some subclasses of `TangentSpace` that make explicit the spaces seen in the uses of TFP so far.

PiperOrigin-RevId: 391643639