Streamline interface for parameter initialization in `make_trainable`.
This removes the `unconstrained_init_fn` argument, in favor of allowing a callable to be passed for the `initial_parameters` argument. Because this callable is given a seed and the constraining bijector, it can implement arbitrary initialization strategies including sampling unconstrained values from a distribution of choice, so this is strictly more general than the previous interface.

PiperOrigin-RevId: 393006350