Add `log_prob` and `log_prob_parts` to `JointDistributionPinned`.
Rewrite docstrings to defend the interpretation thereof as a perfectly
valid non-probability measure on the space of unpinned values.

To justify the existence of `unnormalized_log_prob` and
`unnormalized_log_prob_parts`, mention the possibility of dropping
normalizing constants from those in the future, even though we
currently have no concrete plans to do so.

Still skirting the issue of what measure, exactly, serves as the
base with respect to which `JointDistributionPinned.log_prob` is
a density, since the proposal for that has not yet been implemented.

PiperOrigin-RevId: 395701499