[JAX] Fixes for tensorflow_probability for an upcoming change to jnp.array().
An upcoming change to jax.numpy.array() means that, under a transformation like jax.jit(), it will always stage its arrays into the trace. This often breaks if the array is being used for a shape calculation. Make sure we use static shapes in more places to fix test failures.

PiperOrigin-RevId: 397838341