Unbreak internal code by avoiding extra sample dimension in internal loss computations.
After this change, the sample_shape passed to `surrogate_posterior` is always one-dimensional, and similarly the `target_log_prob_fn` is always called with a single batch dimension. This unbreaks some internal code that relied on these properties. Note that this behavior has never been guaranteed in the contract, and may change in the future, but since this change cleans things up slightly I think it's worth making anyway.

PiperOrigin-RevId: 398152414