Inference Gym: Add the Indiana variant of the Radon model.
The Indiana slice is larger than Minnesota's, and perhaps because of this, the
inference goes a lot smoother in 32 bit precision. I did not verify why.

I had to modify the data loading function to clip floors above 1. This is a NoOp for the Minnesota slice, as that only has floors 0 and 1.

PiperOrigin-RevId: 399253639