Fix tests and experimental/nn/util/convolution_util for numpy 1.20+.
The problem is a change in the behavior of `np.array` when the
argument is a tuple of symbolic Tensors, from auto-stacking to raising
an error.  The upshot for our code is we have to be a little more
careful treating tuples as though they were already stacked Tensors.
Mostly this only comes up in the test suite, and mostly through
`assertAllClose`, which can no longer be relied upon to auto-stack.
Fortunately, we already have `assertAllCloseNested`, which explicitly
takes tuple structure into account, and thus makes for a better test
assertion anyway.

One known problem of this kind remains, namely trying to compute the
shape of an intermediate in Kahan summation, but that fix is less
obvious, so I delay it for a separate change.