Write a shell script that automates setting up a CI emulation environment from scratch.
The script also prints a friendly reminder of the other commands
needed to reproduce a CI test run.

Weakness: this script does not control the Python version, but uses
whatever is on the system.  Is there any way to control said Python
version without using Docker and without altering global state for the
caller?
PiperOrigin-RevId: 402905633