Fix spurious bijector failures when `validate_event_size=True`.
This change updates `_maybe_warn_increased_dof` to always return a list of assertions (which may be the empty list). Previously it would sometimes return `None`, which is not valid as an input to tf.control_dependencies.

This should fix https://github.com/tensorflow/probability/issues/1446.

This also fixes a minor related issue that I discovered while testing the above: the static path gated by `if increased_dof_ is False` was failing to trigger (making it tougher to reproduce the above) because tf.get_static_value returns a `np.bool` scalar rather than the pure Python `False`. The new check `if increased_dof_ is not None and not increased_dof_` does the correct thing in this case.

PiperOrigin-RevId: 404051515