Patch `joint_distribution_sample_path_mixin` to be compatible with numpy 1.20+
The issue was that `np.array` would treat our `Kahan` namedtuple
differently from how it used to work in numpy 1.19, causing spurious
errors in a corner case.

Specifically, if one built a JointDistribution with sample path
semantics, and turned on both experimental_use_kahan_sum and
validate_args at the same time, the JD would try to determine the
shape of a Kahan tuple; in so doing it would try to convert it into a
numpy array, and fail.

Rearranged so JD tests the shapes of the main results of the Kahan
tuples in this scenario.

While digging, also discovered that our shape testing was accidentally
too lax---it would claim that a shape like [3] would add without
broadcasting to a shape like [3, 3], which is not correct.

PiperOrigin-RevId: 404866517